telegram_bot/requirements.txt

python-telegram-bot==20.3
apscheduler==3.10.1
requests==2.31.0
